Fix compilation issue - I"m on OSX, not sure if this was failing only for me. If so, we probably need a way to ensure this works cross plaatform

+++ src/program.ts
@@ -109,7 +109,7 @@ class HeaderFlags {
      headerFlags.str_int16_t_cmp || headerFlags.str_pos || headerFlags.str_len ||
      headerFlags.str_char_code_at || headerFlags.str_substring || headerFlags.str_slice ||
      headerFlags.regex }
-    typedef int int16_t;
+    typedef short int16_t;
 {/if}
 {#if headerFlags.regex}
     struct regex_indices_struct_t {
